исправлены замечания ревью 1
