+++ README.md
@@ -2,3 +2,4 @@
 
 Первый проект на React
 
+Использованы: Redux Toolkit, WebSocket, typeScript
